cleanup modal and make it generic
