+++ src/components/ui/Modal.tsx
@@ -1,33 +1,25 @@
-import { useEffect, useRef } from "react";
+import { useEffect, useRef, ReactNode } from "react";
 import { IoMdClose as CloseIcon } from "react-icons/io";
 import styles from "./Modal.module.css";
 
-export const Modal = ({
-  title,
-  isOpen,
-  onClose,
-}: {
+interface ModalProps {
   title: string;
   isOpen: boolean;
   onClose: () => void;
-}) => {
+  children?: ReactNode[];
+}
+
+export const Modal = ({ title, isOpen, onClose, children }: ModalProps) => {
   const ref = useRef<HTMLDialogElement | null>(null);
 
   useEffect(() => {
-    if (isOpen && !ref.current?.open) {
-      ref.current?.showModal();
-    } else if (!isOpen && ref.current?.open) {
-      close();
-    }
+    if (!isOpen) return;
+    ref.current?.showModal();
+    return close;
   }, [isOpen]);
 
   const close = () => {
     ref.current?.close();
-  };
-
-  const copyToClipboard = async () => {
-    const copyable = window.location.href;
-    await navigator.clipboard.writeText(copyable);
   };
 
   return (
@@ -36,9 +28,7 @@ export const Modal = ({
       <CloseIcon size={25} className={styles.modalCloseIcon} onClick={close} />
       <div className={styles.modalBody}>
         <hr />
-        <p>Share your current game and include the random seed</p>
-        <input value={window.location.href} readOnly />
-        <button onClick={copyToClipboard}>copy</button>
+        {children}
       </div>
     </dialog>
   );
